WB7: split acquisition module and dedupe metadata-core patch fields
Split audible/acquisition into progress, paths, and validation submodules
for clearer ownership; move library probe harness to library_probe.rs.
Extract apply_shared_metadata_patch_fields in abb-metadata-core to unify
apply_to_metadata and to_write_plan field application.

Scout hazards: pub(in audible) visibility preserves sibling imports.
